+++ src/i18n.js
@@ -380,6 +380,8 @@
     messages.ru.list_sort_rating = 'По рейтингу';
     messages.ru.list_sort_year = 'По году выхода';
     messages.ru.list_sort_title = 'По названию А–Я';
+    messages.ru.genre_catalog = 'Выбранный жанр';
+    messages.ru.genre_catalog_fallback = 'Аниме жанра «{genre}»: новые релизы, популярные и завершённые тайтлы.';
     messages.ru.scroll_to_top = 'Наверх';
     messages.ru.catalog_filters = 'Фильтры';
     messages.ru.catalog_filter_reset = 'Сбросить фильтры';
@@ -429,6 +431,8 @@
     messages.en.list_sort_rating = 'By rating';
     messages.en.list_sort_year = 'By release year';
     messages.en.list_sort_title = 'By title A–Z';
+    messages.en.genre_catalog = 'Selected genre';
+    messages.en.genre_catalog_fallback = '{genre} anime: new releases, popular and completed titles.';
     messages.en.scroll_to_top = 'Back to top';
     messages.en.catalog_filters = 'Filters';
     messages.en.catalog_filter_reset = 'Reset filters';
@@ -687,6 +691,8 @@
     messages.uk.list_sort_rating = 'За рейтингом';
     messages.uk.list_sort_year = 'За роком виходу';
     messages.uk.list_sort_title = 'За назвою А–Я';
+    messages.uk.genre_catalog = 'Вибраний жанр';
+    messages.uk.genre_catalog_fallback = 'Аніме жанру «{genre}»: нові релізи, популярні та завершені тайтли.';
     messages.uk.scroll_to_top = 'Нагору';
     messages.uk.catalog_filters = 'Фільтри';
     messages.uk.catalog_filter_reset = 'Скинути фільтри';